refactor(#9): `User`와 `Auth`를 구분짓기 위해 `User`와 관련한 파일을 `user` 패키지로 분리
