--- src/main/java/software_capstone/backend/app/auth/document/Role.java
@@ -1,5 +1,0 @@
-package software_capstone.backend.app.auth.document;
-
-public enum Role {
-    USER, ADMIN
-}
