+++ src/main/java/software_capstone/backend/app/auth/service/AuthService.java
@@ -8,13 +8,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import software_capstone.backend.app.auth.document.OAuthProvider;
 import software_capstone.backend.app.auth.document.RefreshToken;
-import software_capstone.backend.app.auth.document.Role;
-import software_capstone.backend.app.auth.document.User;
+import software_capstone.backend.app.user.document.Role;
+import software_capstone.backend.app.user.document.User;
 import software_capstone.backend.app.auth.dto.LoginRequest;
 import software_capstone.backend.app.auth.dto.TokenResponse;
 import software_capstone.backend.app.auth.jwt.TokenProvider;
 import software_capstone.backend.app.auth.repository.RefreshTokenRepository;
-import software_capstone.backend.app.auth.repository.UserRepository;
+import software_capstone.backend.app.user.repository.UserRepository;
 import software_capstone.backend.global.exception.ErrorMessage;
 import software_capstone.backend.global.exception.UnauthorizedException;
