+++ src/main/java/software_capstone/backend/app/user/document/Role.java
@@ -1,0 +1,5 @@
+package software_capstone.backend.app.user.document;
+
+public enum Role {
+    USER, ADMIN
+}
