+++ src/main/java/software_capstone/backend/app/user/document/User.java
@@ -1,4 +1,4 @@
-package software_capstone.backend.app.auth.document;
+package software_capstone.backend.app.user.document;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.data.mongodb.core.mapping.Document;
 import software_capstone.backend.app.abocado.document.Difficulty;
+import software_capstone.backend.app.auth.document.OAuthProvider;
 import software_capstone.backend.global.document.BaseEntity;
 
 @Getter
